Make code modular
